Force enable autocommit for every new connection
This allows CREATE INDEX CONCURRENTLY to be used in the existing code base, which allows progress to be made on
unit tests. Something somewhere is starting a transaction before it lets us enable autocommit on the Connection,
which then errors out with:
`psycopg.errors.ActiveSqlTransaction: CREATE INDEX CONCURRENTLY cannot run inside a transaction block`
psycopg2 allows this to proceed. I suspect the Twisted ConnectionPool, but haven't found that proof yet