Change a `ViewLike` object to an actual `List`
I'm not certain this is necessary, but seems prudent. From what I understand, a `ViewLike`
object is mutable and could change from underneath us if the database call is retried. I
believe the cost of creating the List is less than the cost of the database accesses
themselves and would be a worthwhile trade-off