Adjust to provide for a generic database engine type
Provide a new key in the database config of "driver" to be specific of the driver to use, which
allows "name" to be a more generic "postgres" allowing swaping of the driver transparently to admins.

Default to "psycopg2", if "name" is set to "postgres". Allow for "name" to remain "psycopg2" for backwards
compatibility.